P6UIC-236 Add parameters to showDialog function
